+++ exts/omni.kit.window.usd_search/docs/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.0.3] - 2024-10-31
+- Only search when search button clicked or enter pressed after input in text field
+- Show host url in result
+
+## [1.0.2] - 2024-10-10
+- Enable search in scene
+- Support proper instance
 
 ## [1.0.1] - 2024-07-31
 - USD Explorer support. Strip deepsearch. Document, comment.